requirements.txt
Django==2.2.6
pytz==2019.3
sqlparse==0.3.0
psycopg2-binary==2.8.4
pip==19.3.1
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
pygments==2.4.2
django-oauth-toolkit==1.2.0
python-dotenv==0.10.3
django-import-export==1.2.0
django-rest-swagger==2.2.0
django-cors-headers==3.2.0
django-rest-passwordreset==1.0.0
numpy==1.17.4
scipy==1.4.0
django-user-activity-log2==0.0.24
python-crontab==2.4
gunicorn==19.9.0
django-health-check